+++ src/lib.rs
@@ -14,11 +14,35 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Signifix. If not, see <https://www.gnu.org/licenses>.
 
-//! Number Formatter of Fixed Significance with Metric Unit Prefix
+//! **Number Formatter of Fixed Significance with Metric Prefix**
 //!
-//! Signifix guarantees a fixed number of significant figures and a fixed number
-//! of resulting characters by automatically choosing the metric unit prefix and
-//! appropriately adjusting the decimal point position.
+//! Formats a given number in one of the two Signifix notations
+//! [as defined below](#definition) by
+//!
+//! 1. selecting the appropriate metric prefix and
+//! 2. moving the decimal point position in a way to sustain a fixed number of
+//!    four significant figures.
+//!
+//! # Definition
+//!
+//! The Signifix notation comprises
+//!
+//!   * a metrically normalized signed significand, that is a decimal of four
+//!     significant figures from `±1.000` to `±999.9` to cover the three powers
+//!     of ten of a particular metric prefix with the three different decimal
+//!     point positions between these four figures, and
+//!   * a metric prefix symbol or its placeholder in case of no prefix
+//!       - either being appended along with a whitespace as in `±1.234␣k`,
+//!         that is the default notation,
+//!       - or replacing the decimal point of the significand as in `±1k234`,
+//!         that is the alternate notation.
+//!
+//! In default notation the placeholder is another whitespace as in `±1.234␣␣`
+//! to align consistently, while in alternate notation it is a number sign as in
+//! `±1#234` to conspicuously separate the integer from the fractional part of
+//! the significand.
+//!
+//! The plus sign of positive numbers is optional.
 //!
 //! # Usage
 //!
@@ -28,7 +52,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! signifix = "0.3.1"
+//! signifix = "0.4.0"
 //! ```
 //!
 //! and this to your crate root:
@@ -41,8 +65,8 @@
 //!
 //! # Examples
 //!
-//! The fixed number of significant figures and resulting characters prevent the
-//! digits and prefixed units from jumping to the left or right:
+//! Both the default and alternate notation result in a fixed number of
+//! characters preventing jumps to the left or right:
 //!
 //! ```
 //! # #![feature(try_from)]
@@ -50,21 +74,22 @@
 //!
 //! use signifix::{Signifix, Result};
 //!
-//! let format = |number| -> Result<String> {
-//! 	Signifix::try_from(number).map(|number| format!("{}", number))
+//! let format = |number| -> Result<(String, String)> {
+//! 	Signifix::try_from(number)
+//! 		.map(|number| (format!("{}", number), format!("{:#}", number)))
 //! };
 //!
-//! assert_eq!(format(1e-04), Ok("100.0 µ".into()));
-//! assert_eq!(format(1e-03), Ok("1.000 m".into()));
-//! assert_eq!(format(1e-02), Ok("10.00 m".into()));
-//! assert_eq!(format(1e-01), Ok("100.0 m".into()));
-//! assert_eq!(format(1e+00), Ok("1.000  ".into()));
-//! assert_eq!(format(1e+01), Ok("10.00  ".into()));
-//! assert_eq!(format(1e+02), Ok("100.0  ".into()));
-//! assert_eq!(format(1e+03), Ok("1.000 k".into()));
-//! assert_eq!(format(1e+04), Ok("10.00 k".into()));
-//! assert_eq!(format(1e+05), Ok("100.0 k".into()));
-//! assert_eq!(format(1e+06), Ok("1.000 M".into()));
+//! assert_eq!(format(1e-04), Ok(("100.0 µ".into(), "100µ0".into())));
+//! assert_eq!(format(1e-03), Ok(("1.000 m".into(), "1m000".into())));
+//! assert_eq!(format(1e-02), Ok(("10.00 m".into(), "10m00".into())));
+//! assert_eq!(format(1e-01), Ok(("100.0 m".into(), "100m0".into())));
+//! assert_eq!(format(1e+00), Ok(("1.000  ".into(), "1#000".into())));
+//! assert_eq!(format(1e+01), Ok(("10.00  ".into(), "10#00".into())));
+//! assert_eq!(format(1e+02), Ok(("100.0  ".into(), "100#0".into())));
+//! assert_eq!(format(1e+03), Ok(("1.000 k".into(), "1k000".into())));
+//! assert_eq!(format(1e+04), Ok(("10.00 k".into(), "10k00".into())));
+//! assert_eq!(format(1e+05), Ok(("100.0 k".into(), "100k0".into())));
+//! assert_eq!(format(1e+06), Ok(("1.000 M".into(), "1M000".into())));
 //! ```
 //!
 //! This is useful to smoothly refresh a transfer rate within a terminal:
@@ -76,7 +101,7 @@
 //! use std::convert::TryFrom; // Until stabilized.
 //!
 //! use std::f64;
-//! use signifix::{Signifix, Error, MIN_LEN};
+//! use signifix::{Signifix, Error, DEF_MIN_LEN};
 //!
 //! let format_rate = |bytes: u128, nanoseconds: u128| -> String {
 //! 	let bytes_per_second = bytes as f64 / nanoseconds as f64 * 1e+09;
@@ -101,7 +126,7 @@
 //! 			Error::Nan => " - ---- - ", // zero bytes in zero nanoseconds
 //! 		}.into(),
 //! 	};
-//! 	assert_eq!(rate.chars().count(), MIN_LEN + unit.chars().count());
+//! 	assert_eq!(rate.chars().count(), DEF_MIN_LEN + unit.chars().count());
 //! 	rate
 //! };
 //!
@@ -114,17 +139,18 @@
 //! ```
 //!
 //! Or to monitor a measured quantity like an electrical current including its
-//! direction with an optional space placeholder to align with negative values:
+//! direction with positive numbers being padded to align with negative ones:
 //!
 //! ```
 //! # #![feature(try_from)]
 //! use std::convert::TryFrom; // Until stabilized.
 //!
-//! use signifix::{Signifix, Result, Error};
+//! use signifix::{Signifix, Result, DEF_MAX_LEN};
 //!
-//! let format_load = |current| -> Result<String> {
-//! 	if let Some(current) = current {
-//! 		Signifix::try_from(current).map(|current| format!("{:#}A", current))
+//! let format_load = |amps| -> Result<String> {
+//! 	if let Some(amps) = amps {
+//! 		Signifix::try_from(amps)
+//! 			.map(|amps| format!("{:>1$}A", amps, DEF_MAX_LEN))
 //! 	} else {
 //! 		Ok("     0  A".into())
 //! 	}
@@ -135,8 +161,8 @@
 //! assert_eq!(format_load(Some(-2.927e-06)), Ok("-2.927 µA".into()));
 //! ```
 //!
-//! In case of displaying a file size difference after modification, a plus sign
-//! might be preferred for positive values:
+//! While to visualize a change in file size, a plus sign might be preferred for
+//! positive numbers:
 //!
 //! ```
 //! # #![feature(try_from)]
@@ -145,14 +171,14 @@
 //! use signifix::{Signifix, Result, Error};
 //!
 //! let format_diff = |curr, prev| -> Result<String> {
-//! 	Signifix::try_from(curr - prev).map(|diff| format!("{:+}B", diff))
+//! 	Signifix::try_from(curr - prev).map(|diff| format!("{:+#}", diff))
 //! 		.or_else(|case| if case == Error::OutOfLowerBound(0f64)
-//! 			{ Ok("     0  B".into()) } else { Err(case) })
+//! 			{ Ok("=const".into()) } else { Err(case) })
 //! };
 //!
-//! assert_eq!(format_diff(78_346, 57_393), Ok("+20.95 kB".into()));
-//! assert_eq!(format_diff(93_837, 93_837), Ok("     0  B".into()));
-//! assert_eq!(format_diff(27_473, 46_839), Ok("-19.37 kB".into()));
+//! assert_eq!(format_diff(78_346, 57_393), Ok("+20k95".into()));
+//! assert_eq!(format_diff(93_837, 93_837), Ok("=const".into()));
+//! assert_eq!(format_diff(27_473, 36_839), Ok("-9k366".into()));
 //! ```
 
 #![deny(missing_docs)]
@@ -175,15 +201,15 @@ use Ordering::Equal;
 /// `Signifix` type.
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub enum Error {
-	/// The given number is below the lower bound ±1.000 y (= ±1e-24) of the
-	/// lowermost metric unit prefix yocto (y = 1e-24).
+	/// The given number is below the lower bound `±1.000 y` (`= ±1e-24`) of the
+	/// lowermost metric prefix yocto (`y = 1e-24`).
 	OutOfLowerBound(f64),
 
-	/// The given number is above the upper bound ±999.9 Y (≅ ±1e+27) of the
-	/// uppermost metric unit prefix yotta (Y = 1e+24).
+	/// The given number is above the upper bound `±999.9 Y` (`≅ ±1e+27`) of the
+	/// uppermost metric prefix yotta (`Y = 1e+24`).
 	OutOfUpperBound(f64),
 
-	/// The given number is actually Not a Number (NaN).
+	/// The given number is actually not a number (NaN).
 	Nan,
 }
 
@@ -229,15 +255,17 @@ impl error::Error for Error {
 pub type Result<T> = result::Result<T, Error>;
 
 /// Intermediate implementor type of this crate's `TryFrom` and `Display` trait
-/// implementations. Former tries to convert a number to this type by
-/// automatically choosing the metric unit prefix and appropriately adjusting
+/// implementations. Former tries to convert a given number to this type by
+/// calculating the appropriate metric prefix, the normalized significand, and
 /// the decimal point position while latter uses this type's fields to format
-/// the number to a string of a fixed number of significant figures and a fixed
-/// number of resulting characters. The format parameter `{}` prefixes negative
-/// numbers with a minus sign. The variant `{:+}` additionally prefixes positive
-/// numbers with a plus sign while the variant `{:#}` uses a whitespace instead
-/// to align with negative numbers. Padding via fill/alignment formatting
-/// parameters like `{:>8}` is supported as well.
+/// the number to a string of four significant figures inclusive the metric
+/// prefix symbol.
+///
+/// Interpreted formatting parameters are
+///
+/// * `#` to indicate the alternate notation,
+/// * `+` to prefix positive numbers with a plus sign,
+/// * `fill`, `alignment`, and `width` to pad or align numbers.
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub struct Signifix {
 	numerator: i16,
@@ -266,60 +294,94 @@ impl Ord for Signifix {
 	}
 }
 
-/// Number of resulting characters when no sign or whitespace is prefixed.
-pub const MIN_LEN: usize = 7;
+/// Number of resulting characters in default notation when no sign or
+/// whitespace is prefixed.
+pub const DEF_MIN_LEN: usize = 7;
 
-/// Number of resulting characters when a sign or whitespace is prefixed.
-pub const MAX_LEN: usize = 8;
+/// Number of resulting characters in default notation when a sign or
+/// whitespace is prefixed.
+pub const DEF_MAX_LEN: usize = 8;
 
-/// Metric unit prefix symbols from y to Y indexed from 0 to 16.
-pub const SYMBOL: [char; 17] = [
-	'y', 'z', 'a', 'f', 'p', 'n', 'µ', 'm',
-	' ',
-	'k', 'M', 'G', 'T', 'P', 'E', 'Z', 'Y',
+/// Number of resulting characters in alternate notation when no sign or
+/// whitespace is prefixed.
+pub const ALT_MIN_LEN: usize = 5;
+
+/// Number of resulting characters in alternate notation when a sign or
+/// whitespace is prefixed.
+pub const ALT_MAX_LEN: usize = 6;
+
+/// Metric prefix symbols from `Some('y')` to `Some('m')` indexed from `0` to
+/// `7` and from `Some('k')` to `Some('Y')` indexed from `9` to `16`, or `None`
+/// indexed at `8`.
+pub const SYMBOLS: [Option<char>; 17] = [
+	Some('y'), Some('z'), Some('a'), Some('f'),
+	Some('p'), Some('n'), Some('µ'), Some('m'),
+	None,
+	Some('k'), Some('M'), Some('G'), Some('T'),
+	Some('P'), Some('E'), Some('Z'), Some('Y'),
 ];
 
-/// Metric unit prefix factors from 1e-24 to 1e+24 indexed from 0 to 16.
-pub const FACTOR: [f64; 17] = [
-	1e-24, 1e-21, 1e-18, 1e-15, 1e-12, 1e-09, 1e-06, 1e-03,
+/// Metric prefix factors from `1e-24` to `1e-03` indexed from `0` to `7` and
+/// from `1e+03` to `1e+24` indexed from `9` to `16`, or `1e+00` indexed at `8`.
+pub const FACTORS: [f64; 17] = [
+	1e-24, 1e-21, 1e-18, 1e-15,
+	1e-12, 1e-09, 1e-06, 1e-03,
 	1e+00,
-	1e+03, 1e+06, 1e+09, 1e+12, 1e+15, 1e+18, 1e+21, 1e+24,
+	1e+03, 1e+06, 1e+09, 1e+12,
+	1e+15, 1e+18, 1e+21, 1e+24,
 ];
 
 impl Signifix {
-	/// Metrically normalized signed significand from ±1.000 to ±999.9.
+	/// Metrically normalized signed significand from `±1.000` to `±999.9`.
 	pub fn significand(&self) -> f64 {
 		self.numerator() as f64 / self.denominator() as f64
 	}
 
-	/// Signed significand numerator from ±1,000 to ±9,999.
+	/// Signed significand numerator from `±1,000` to `±9,999`.
 	pub fn numerator(&self) -> i32 {
 		self.numerator.into()
 	}
 
-	/// Significand denominator of either 10, 100, or 1,000.
+	/// Significand denominator of either `10`, `100`, or `1,000`.
 	pub fn denominator(&self) -> i32 {
 		[1, 10, 100, 1_000][self.exponent()]
 	}
 
-	/// Exponent of significand denominator of either 1, 2, or 3.
+	/// Exponent of significand denominator of either `1`, `2`, or `3`.
 	pub fn exponent(&self) -> usize {
 		self.exponent.into()
 	}
 
-	/// Metric unit prefix as `SYMBOL` and `FACTOR` array index from 0 to 16.
+	/// Signed integer part of significand from `±1` to `±999`.
+	pub fn integer(&self) -> i32 {
+		self.numerator() / self.denominator()
+	}
+
+	/// Fractional part of significand from `0` to `999`.
+	pub fn fractional(&self) -> i32 {
+		self.numerator().abs() % self.denominator()
+	}
+
+	/// Signed integer and fractional part at once, in given order.
+	pub fn parts(&self) -> (i32, i32) {
+		let trunc = self.numerator() / self.denominator();
+		let fract = self.numerator() - self.denominator() * trunc;
+		(trunc, fract.abs())
+	}
+
+	/// Metric prefix as `SYMBOLS` and `FACTORS` array index from `0` to `16`.
 	pub fn prefix(&self) -> usize {
 		self.prefix.into()
 	}
 
-	/// Symbol of metric unit prefix from y to Y.
-	pub fn symbol(&self) -> char {
-		SYMBOL[self.prefix()]
+	/// Symbol of metric prefix from `Some('y')` to `Some('Y')`, or `None`.
+	pub fn symbol(&self) -> Option<char> {
+		SYMBOLS[self.prefix()]
 	}
 
-	/// Factor of metric unit prefix from 1e-24 to 1e+24.
+	/// Factor of metric prefix from `1e-24` to `1e+24`, or `1e+00`.
 	pub fn factor(&self) -> f64 {
-		FACTOR[self.prefix()]
+		FACTORS[self.prefix()]
 	}
 }
 
@@ -348,39 +410,39 @@ impl TryFrom<f64> for Signifix {
 	fn try_from(number: f64) -> Result<Self> {
 		let (numerator, prefix) = {
 			let number = number.abs();
-			if number < FACTOR[08] {
-				let prefix = if number < FACTOR[04] {
-					if number < FACTOR[02] {
-						if number < FACTOR[01] { 00 } else { 01 }
+			if number < FACTORS[08] {
+				let prefix = if number < FACTORS[04] {
+					if number < FACTORS[02] {
+						if number < FACTORS[01] { 00 } else { 01 }
 					} else {
-						if number < FACTOR[03] { 02 } else { 03 }
+						if number < FACTORS[03] { 02 } else { 03 }
 					}
 				} else {
-					if number < FACTOR[06] {
-						if number < FACTOR[05] { 04 } else { 05 }
+					if number < FACTORS[06] {
+						if number < FACTORS[05] { 04 } else { 05 }
 					} else {
-						if number < FACTOR[07] { 06 } else { 07 }
+						if number < FACTORS[07] { 06 } else { 07 }
 					}
 				};
-				(number / FACTOR[prefix], prefix)
+				(number / FACTORS[prefix], prefix)
 			} else
-			if number < FACTOR[09] {
+			if number < FACTORS[09] {
 				(number, 08)
 			} else {
-				let prefix = if number < FACTOR[13] {
-					if number < FACTOR[11] {
-						if number < FACTOR[10] { 09 } else { 10 }
+				let prefix = if number < FACTORS[13] {
+					if number < FACTORS[11] {
+						if number < FACTORS[10] { 09 } else { 10 }
 					} else {
-						if number < FACTOR[12] { 11 } else { 12 }
+						if number < FACTORS[12] { 11 } else { 12 }
 					}
 				} else {
-					if number < FACTOR[15] {
-						if number < FACTOR[14] { 13 } else { 14 }
+					if number < FACTORS[15] {
+						if number < FACTORS[14] { 13 } else { 14 }
 					} else {
-						if number < FACTOR[16] { 15 } else { 16 }
+						if number < FACTORS[16] { 15 } else { 16 }
 					}
 				};
-				(number / FACTOR[prefix], prefix)
+				(number / FACTORS[prefix], prefix)
 			}
 		};
 		let scaled = |pow: f64| {
@@ -419,7 +481,7 @@ impl TryFrom<f64> for Signifix {
 				})
 			} else {
 				let prefix = prefix + 1;
-				if prefix < FACTOR.len() {
+				if prefix < FACTORS.len() {
 					Ok(Signifix {
 						numerator: signed(1e+03) as i16,
 						exponent: 3,
@@ -439,16 +501,25 @@ impl TryFrom<f64> for Signifix {
 
 impl Display for Signifix {
 	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-		if f.sign_plus() {
-			f.pad(&format!("{:+.*} {}",
-				self.exponent(), self.significand(), self.symbol()))
-		} else
-		if f.alternate() && self.numerator().is_positive() {
-			f.pad(&format!(" {:.*} {}",
-				self.exponent(), self.significand(), self.symbol()))
+		if f.alternate() {
+			let symbol = self.symbol().unwrap_or('#');
+			let (integer, fractional) = self.parts();
+			if f.sign_plus() {
+				f.pad(&format!("{:+}{}{:03$}",
+					integer, symbol, fractional, self.exponent()))
+			} else {
+				f.pad(&format!("{:}{}{:03$}",
+					integer, symbol, fractional, self.exponent()))
+			}
 		} else {
-			f.pad(&format!("{:.*} {}",
-				self.exponent(), self.significand(), self.symbol()))
+			let symbol = self.symbol().unwrap_or(' ');
+			if f.sign_plus() {
+				f.pad(&format!("{:+.*} {}",
+					self.exponent(), self.significand(), symbol))
+			} else {
+				f.pad(&format!("{:.*} {}",
+					self.exponent(), self.significand(), symbol))
+			}
 		}
 	}
 }
@@ -459,79 +530,90 @@ mod tests {
 	use std::f64;
 	use std::mem::size_of;
 
-	fn def(number: f64) -> Result<String> {
-		Signifix::try_from(number).map(|number| format!("{}", number))
+	fn fmt(number: f64) -> Result<(String, String)> {
+		Signifix::try_from(number).map(|number| (
+			format!("{}",   number),
+			format!("{:#}", number),
+		))
 	}
-	fn pos(number: f64) -> Result<String> {
-		Signifix::try_from(number).map(|number| format!("{:+}", number))
+	fn pos(number: f64) -> Result<(String, String)> {
+		Signifix::try_from(number).map(|number| (
+			format!("{:+}",  number),
+			format!("{:+#}", number),
+		))
 	}
-	fn alt(number: f64) -> Result<String> {
-		Signifix::try_from(number).map(|number| format!("{:#}", number))
+	fn pad(number: f64) -> Result<(String, String)> {
+		Signifix::try_from(number).map(|number| (
+			format!("{:>1$}",  number, DEF_MAX_LEN),
+			format!("{:>#1$}", number, ALT_MAX_LEN),
+		))
 	}
 
 	#[test]
-	fn metric_unit_prefix() {
-		assert_eq!(def(1e-24), Ok("1.000 y".into()));
-		assert_eq!(def(1e-21), Ok("1.000 z".into()));
-		assert_eq!(def(1e-18), Ok("1.000 a".into()));
-		assert_eq!(def(1e-15), Ok("1.000 f".into()));
-		assert_eq!(def(1e-12), Ok("1.000 p".into()));
-		assert_eq!(def(1e-09), Ok("1.000 n".into()));
-		assert_eq!(def(1e-06), Ok("1.000 µ".into()));
-		assert_eq!(def(1e-03), Ok("1.000 m".into()));
-		assert_eq!(def(1e+00), Ok("1.000  ".into()));
-		assert_eq!(def(1e+03), Ok("1.000 k".into()));
-		assert_eq!(def(1e+06), Ok("1.000 M".into()));
-		assert_eq!(def(1e+09), Ok("1.000 G".into()));
-		assert_eq!(def(1e+12), Ok("1.000 T".into()));
-		assert_eq!(def(1e+15), Ok("1.000 P".into()));
-		assert_eq!(def(1e+18), Ok("1.000 E".into()));
-		assert_eq!(def(1e+21), Ok("1.000 Z".into()));
-		assert_eq!(def(1e+24), Ok("1.000 Y".into()));
+	fn factors_to_symbols() {
+		assert_eq!(fmt(1e-24), Ok(("1.000 y".into(), "1y000".into())));
+		assert_eq!(fmt(1e-21), Ok(("1.000 z".into(), "1z000".into())));
+		assert_eq!(fmt(1e-18), Ok(("1.000 a".into(), "1a000".into())));
+		assert_eq!(fmt(1e-15), Ok(("1.000 f".into(), "1f000".into())));
+		assert_eq!(fmt(1e-12), Ok(("1.000 p".into(), "1p000".into())));
+		assert_eq!(fmt(1e-09), Ok(("1.000 n".into(), "1n000".into())));
+		assert_eq!(fmt(1e-06), Ok(("1.000 µ".into(), "1µ000".into())));
+		assert_eq!(fmt(1e-03), Ok(("1.000 m".into(), "1m000".into())));
+		assert_eq!(fmt(1e+00), Ok(("1.000  ".into(), "1#000".into())));
+		assert_eq!(fmt(1e+03), Ok(("1.000 k".into(), "1k000".into())));
+		assert_eq!(fmt(1e+06), Ok(("1.000 M".into(), "1M000".into())));
+		assert_eq!(fmt(1e+09), Ok(("1.000 G".into(), "1G000".into())));
+		assert_eq!(fmt(1e+12), Ok(("1.000 T".into(), "1T000".into())));
+		assert_eq!(fmt(1e+15), Ok(("1.000 P".into(), "1P000".into())));
+		assert_eq!(fmt(1e+18), Ok(("1.000 E".into(), "1E000".into())));
+		assert_eq!(fmt(1e+21), Ok(("1.000 Z".into(), "1Z000".into())));
+		assert_eq!(fmt(1e+24), Ok(("1.000 Y".into(), "1Y000".into())));
 	}
 	#[test]
 	fn fixed_significance() {
-		assert_eq!(def(1.234e+02), Ok("123.4  ".into()));
-		assert_eq!(def(1.234e+03), Ok("1.234 k".into()));
-		assert_eq!(def(1.234e+04), Ok("12.34 k".into()));
-		assert_eq!(def(1.234e+05), Ok("123.4 k".into()));
-		assert_eq!(def(1.234e+06), Ok("1.234 M".into()));
+		assert_eq!(fmt(1.234e+02), Ok(("123.4  ".into(), "123#4".into())));
+		assert_eq!(fmt(1.234e+03), Ok(("1.234 k".into(), "1k234".into())));
+		assert_eq!(fmt(1.234e+04), Ok(("12.34 k".into(), "12k34".into())));
+		assert_eq!(fmt(1.234e+05), Ok(("123.4 k".into(), "123k4".into())));
+		assert_eq!(fmt(1.234e+06), Ok(("1.234 M".into(), "1M234".into())));
 	}
 	#[test]
 	fn formatting_options() {
-		assert_eq!(def(-1e+00), Ok("-1.000  ".into()));
-		assert_eq!(def( 1e+00), Ok( "1.000  ".into()));
-		assert_eq!(pos(-1e+00), Ok("-1.000  ".into()));
-		assert_eq!(pos( 1e+00), Ok("+1.000  ".into()));
-		assert_eq!(alt(-1e+00), Ok("-1.000  ".into()));
-		assert_eq!(alt( 1e+00), Ok(" 1.000  ".into()));
+		assert_eq!(fmt(-1e+00), Ok(("-1.000  ".into(), "-1#000".into())));
+		assert_eq!(fmt( 1e+00), Ok(( "1.000  ".into(),  "1#000".into())));
+		assert_eq!(pos(-1e+00), Ok(("-1.000  ".into(), "-1#000".into())));
+		assert_eq!(pos( 1e+00), Ok(("+1.000  ".into(), "+1#000".into())));
+		assert_eq!(pad(-1e+00), Ok(("-1.000  ".into(), "-1#000".into())));
+		assert_eq!(pad( 1e+00), Ok((" 1.000  ".into(), " 1#000".into())));
 	}
 	#[test]
 	fn lower_prefix_bound() {
-		assert_eq!(def(-0.99951e-24), Ok("-1.000 y".into()));
-		assert_eq!(def(-0.99949e-24),
+		assert_eq!(fmt(-0.99951e-24),
+			Ok(("-1.000 y".into(), "-1y000".into())));
+		assert_eq!(fmt(-0.99949e-24),
 			Err(Error::OutOfLowerBound(-0.99949e-24)));
 	}
 	#[test]
 	fn upper_prefix_bound() {
-		assert_eq!(def(-0.999949e+27), Ok("-999.9 Y".into()));
-		assert_eq!(def(-0.999951e+27),
+		assert_eq!(fmt(-0.999949e+27),
+			Ok(("-999.9 Y".into(), "-999Y9".into())));
+		assert_eq!(fmt(-0.999951e+27),
 			Err(Error::OutOfUpperBound(-0.999951e+27)));
 	}
 	#[test]
 	fn upper_prefix_round() {
-		assert_eq!(def(0.999949e+06), Ok("999.9 k".into()));
-		assert_eq!(def(0.999951e+06), Ok("1.000 M".into()));
+		assert_eq!(fmt(0.999949e+06), Ok(("999.9 k".into(), "999k9".into())));
+		assert_eq!(fmt(0.999951e+06), Ok(("1.000 M".into(), "1M000".into())));
 	}
 	#[test]
 	fn fp_category_safety() {
-		assert_eq!(def(0f64),
+		assert_eq!(fmt(0f64),
 			Err(Error::OutOfLowerBound(0f64)));
-		assert_eq!(def(f64::NEG_INFINITY),
+		assert_eq!(fmt(f64::NEG_INFINITY),
 			Err(Error::OutOfUpperBound(f64::NEG_INFINITY)));
-		assert_eq!(def(f64::INFINITY),
+		assert_eq!(fmt(f64::INFINITY),
 			Err(Error::OutOfUpperBound(f64::INFINITY)));
-		assert_eq!(def(f64::NAN),
+		assert_eq!(fmt(f64::NAN),
 			Err(Error::Nan));
 	}
 	#[test]